+++ stream-chat-android-state/src/test/java/io/getstream/chat/android/state/plugin/listener/internal/DeleteMessageListenerStateTest.kt
@@ -42,7 +42,7 @@ internal class DeleteMessageListenerStateTest {
 
     private val channelStateLogic: ChannelStateLogic = mock()
     private val channelLogic: ChannelLogic = mock {
-        on(it.stateLogic()) doReturn channelStateLogic
+        on(it.stateLogic) doReturn channelStateLogic
     }
     private val threadsLogic: QueryThreadsLogic = mock()
     private val activeThreadsLogic = listOf(threadsLogic)